check in refactored replace_macaddress_with_assetuid and the unit test
