+++ modules/replace_macaddress_with_assetuid/tests/pfs/MAP_PATH/li7200/li7200_stcdd/001C940419DC/li7200_stcdd_001C940419DC.txt
@@ -1,0 +1,1 @@
+27639
